requirements.txt
facebook-business>=19.0.0
supabase>=2.0.0
anthropic>=0.40.0
click>=8.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
tabulate>=0.9.0
rich>=13.0.0
resend>=2.0.0
